moved functions viewCampaigns and uploadCampaign to campaign.js module and added sequence diagrams #64 #65 #66
