+++ lib/campaign.js
@@ -15,4 +15,14 @@ exports.uploadCampaign = function (title,budget,tags,advertiserId){
     });
 }
 
+exports.viewCampaigns = function (advertiserID) {
+  var deferred = Q.defer();
+
+  Campaign.findAll({where : { advertiserId : advertiserID } }).then(function (result) {
+    deferred.resolve(result);
+    
+  });
+  return deferred.promise;
+}
+
